// speedyGo-ui/src/app/services/models/page-response-comment.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Comment } from '../models/comment';
export interface PageResponseComment {
  content?: Array<Comment>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}